refactor: 오류 처리
